Refactor file names of stx hook class
